+++ C/palindrome.c
