+++ Java/Push Zeros to end.java
